tests/test-requirements.txt
dnspython==1.12.0
oslo.config==2.4.0
flake8==2.4.1
